Rename regexTypes.ts to TokenizerOptions.ts
